backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
torch==2.0.1
torchaudio==2.0.2
pytorch-lightning==2.0.9
pyannote.audio==3.1.1
pydub==0.25.1
python-dotenv==1.0.0
soundfile==0.12.1
requests==2.31.0
ffmpeg-python==0.2.0